fix num_api_keys was compact array not array
not sure how this was supposed to be known, but turns out that kafka has
json schema definitions for request and responses in the codebase, where
versions are marked as "flexible", which means compact arrays are used
instead of normal arrays. e.g. schema for apiversions request is here:
https://github.com/apache/kafka/blob/trunk/clients/src/main/resources/common/message/ApiVersionsRequest.json
discovered this from this blog: https://ivanyu.me/blog/2024/09/08/kafka-protocol-practical-guide/

unsigned varint is not defined in the protocol doc, but it's the same as
varint, as varints don't support signed integers. This is specified in
the protobuf doc: https://protobuf.dev/programming-guides/encoding/#varints