serial requests
need to disable netcat based tests as netcat doesn't know when the
server is done responding: we need to kill the server manually to make
it flush the remaining contents. attempted killing the thread but it was
too much overhead, so just disabling for now, as we already have rust
based tests that are verifying almost the same things.